Fixes some issues with the control list rename
