+++ commec/tests/test_regulation.py
@@ -2,19 +2,20 @@ import logging
 import pytest
 import pandas as pd
 
-import commec.control_list.containers as crc
+from commec.control_list.containers import AccessionFormat, derive_accession_type, RegulationList, ListMode
+import commec.control_list.list_data as ld
 from commec.control_list.control_list import get_regulation, post_process_regulation_data
 
 from commec.utils.logger import setup_console_logging
 
 @pytest.mark.parametrize("new_list,expected_outcome", [
     pytest.param(*case) for case in [
-        (crc.RegulationList("list01","L1","www.l1.com",["NZ"],1), True), # Duplicate
-        (crc.RegulationList("list01","L2","www.l1.com",["NZ"],0), True), # Non-Duplicate
-        (crc.RegulationList("list01","L1","www.l2.com",["NZ"],0), False), # Wrong URL
-        (crc.RegulationList("list01","L1","www.l1.com",["NZ", "AU"],0), False), # Wrong Regions
-        (crc.RegulationList("list01","L1","www.l1.com",["AU"],0), False), # Wrong Region
-        (crc.RegulationList("list02","L1","www.l1.com",["NZ"],0), False), # Wrong Name
+        (RegulationList("list01","L1","www.l1.com",["NZ"],1), True), # Duplicate
+        (RegulationList("list01","L2","www.l1.com",["NZ"],0), True), # Non-Duplicate
+        (RegulationList("list01","L1","www.l2.com",["NZ"],0), False), # Wrong URL
+        (RegulationList("list01","L1","www.l1.com",["NZ", "AU"],0), False), # Wrong Regions
+        (RegulationList("list01","L1","www.l1.com",["AU"],0), False), # Wrong Region
+        (RegulationList("list02","L1","www.l1.com",["NZ"],0), False), # Wrong Name
     ]
 ])
 def test_list_overwrite_protections(new_list, expected_outcome):
@@ -22,18 +23,18 @@ def test_list_overwrite_protections(new_list, expected_outcome):
     Tests what occurs when two lists share an acronym that should not be concatenated.
     """
     setup_console_logging(logging.DEBUG)
-    existing_list = crc.RegulationList("list01","L1","www.l1.com",["NZ"],0)
-    crc.add_regulated_list(existing_list)
-    assert expected_outcome == crc.add_regulated_list(new_list)
+    existing_list = RegulationList("list01","L1","www.l1.com",["NZ"],0)
+    ld.add_regulated_list(existing_list)
+    assert expected_outcome == ld.add_regulated_list(new_list)
 
 def test_multiple_entrys():
     """
     Outcomes for where multiple lists contain duplicate taxid entries.
     """
-    crc.clear()
+    ld.clear()
     setup_console_logging(logging.DEBUG)
-    assert crc.add_regulated_list(crc.RegulationList("List01","L1","www.list1.com",["NZ"],crc.ListMode.COMPLIANCE))
-    assert crc.add_regulated_list(crc.RegulationList("List02","L2","www.list2.com",["AU"],crc.ListMode.COMPLIANCE))
+    assert ld.add_regulated_list(RegulationList("List01","L1","www.list1.com",["NZ"],ListMode.COMPLIANCE))
+    assert ld.add_regulated_list(RegulationList("List02","L2","www.list2.com",["AU"],ListMode.COMPLIANCE))
 
     input_list1_data = pd.DataFrame([
         {
@@ -100,27 +101,27 @@ def test_multiple_entrys():
     print(input_list2_data.to_string())
 
     print("adding data...")
-    crc.add_regulated_taxid_data(input_list1_data)
-    crc.add_regulated_taxid_data(input_list2_data)
+    ld.add_regulated_taxid_data(input_list1_data)
+    ld.add_regulated_taxid_data(input_list2_data)
 
     print("Preprocessing:")
-    print(crc.REGULATED_TAXID_ANNOTATIONS.to_string())
+    print(ld.REGULATED_TAXID_ANNOTATIONS.to_string())
     post_process_regulation_data()
     print("Postprocessing:")
-    print(crc.REGULATED_TAXID_ANNOTATIONS.to_string())
-    print(crc.REGULATED_TAXID_ANNOTATIONS.shape)
+    print(ld.REGULATED_TAXID_ANNOTATIONS.to_string())
+    print(ld.REGULATED_TAXID_ANNOTATIONS.shape)
 
     # Twelve headings, however should only have 3 entries from the above 5 entries.
-    assert crc.REGULATED_TAXID_ANNOTATIONS.shape == (3,12), "Incorrect number of imported Regulation Annotations."
+    assert ld.REGULATED_TAXID_ANNOTATIONS.shape == (3,12), "Incorrect number of imported Regulation Annotations."
 
-    output = get_regulation(11320, crc.AccessionFormat.TAXID)
+    output = get_regulation(11320, AccessionFormat.TAXID)
     assert len(output) == 2, "Incorrect number of returned Regulations."
 
 @pytest.mark.parametrize("accessions,expected_outcome", [
     pytest.param(*case) for case in [
-        (["111","4","11084", 444], crc.AccessionFormat.TAXID),
-        (["HG992755.1","CAG2243592.1", "CP001814.1", "DI192294.1"], crc.AccessionFormat.GENBANK),
-        (["Q5VW38", "Q7L1I2", "V5XZS6", "B1P1E1"], crc.AccessionFormat.UNIPROT),
+        (["111","4","11084", 444], AccessionFormat.TAXID),
+        (["HG992755.1","CAG2243592.1", "CP001814.1", "DI192294.1"], AccessionFormat.GENBANK),
+        (["Q5VW38", "Q7L1I2", "V5XZS6", "B1P1E1"], AccessionFormat.UNIPROT),
     ]
 ])
 def test_accession_identification_format(accessions, expected_outcome):
@@ -131,5 +132,5 @@ def test_accession_identification_format(accessions, expected_outcome):
     """
     setup_console_logging(logging.DEBUG)
     for accession in accessions:
-        outcome = crc.derive_accession_type(accession)
+        outcome = derive_accession_type(accession)
         assert expected_outcome == outcome, f"{accession} failed to be identified. Expected {expected_outcome}, got {outcome}"
